feat: made some changes and updated resume
