+++ app/components/Loader.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Loader = () => {
+  return (
+    <div>Loader</div>
+  )
+}
+
+export default Loader
