+++ app/lib/TypeWriter.tsx
