+++ app/lib/data.ts
@@ -110,3 +110,21 @@ export const projects = [
     ],
   },
 ];
+
+export const icon_Links = [
+  {
+    link: "https://twitter.com/joshi__utsav",
+    alt: "twitter",
+    src: "/twitterWhite.png",
+  },
+  {
+    link: "https://github.com/joshiUtsav",
+    alt: "github",
+    src: "/githubwhite.png",
+  },
+  {
+    link: "https://www.linkedin.com/in/joshi-utsav/",
+    alt: "linkedIn",
+    src: "/linkedinwhite.png",
+  },
+];
